S3/requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
uvicorn==0.15.0
Pillow==8.3.2
numpy==1.21.2
opencv-python==4.5.3.56
jinja2==3.0.1
aiofiles==0.8.0 